[Refactor]: 컬럼스크롤 로직 분리, 디폴트 프로필이미지 변경
